Added installed scripts control

+++ basic/dotnet.sh
@@ -1,8 +1,25 @@
 #!/bin/bash
 
-# .NET Core
-wget https://packages.microsoft.com/config/debian/10/packages-microsoft-prod.deb -4 -O packages-microsoft-prod.deb
-sudo dpkg -i packages-microsoft-prod.deb
-sudo apt update
-sudo apt install -y dotnet-sdk-6.0
-rm packages-microsoft-prod.deb
+INSTALL_NAME="Dotnet"
+
+INSTALL_FILE="$HOME/.config/linux-installer"
+INSTALL_ALREADY=false
+if [ -f "$INSTALL_FILE" ]; then
+    INSTALLED_NAME=`cat $INSTALL_FILE | grep $INSTALL_NAME`
+    [[ $INSTALLED_NAME == $INSTALL_NAME ]] && INSTALL_ALREADY=true
+fi
+
+if [ "$INSTALL_ALREADY" == true ]; then
+    echo "$INSTALL_NAME is already installed."
+else
+    echo "Installing $INSTALL_NAME."
+
+    wget https://packages.microsoft.com/config/debian/10/packages-microsoft-prod.deb -4 -O packages-microsoft-prod.deb
+    sudo dpkg -i packages-microsoft-prod.deb
+    sudo apt update
+    sudo apt install -y dotnet-sdk-6.0
+    rm packages-microsoft-prod.deb
+
+    [[ ! -d "$HOME/.config" ]] && mkdir $HOME/.config
+    echo $INSTALL_NAME >> $INSTALL_FILE
+fi
