+++ virtual/helm.sh
@@ -1,4 +1,24 @@
-curl https://baltocdn.com/helm/signing.asc | sudo apt-key add -
-echo "deb https://baltocdn.com/helm/stable/debian/ all main" | sudo tee /etc/apt/sources.list.d/helm-stable-debian.list
-sudo apt-get update
-sudo apt-get install -y helm
+#!/bin/bash
+
+INSTALL_NAME="Helm"
+
+INSTALL_FILE="$HOME/.config/linux-installer"
+INSTALL_ALREADY=false
+if [ -f "$INSTALL_FILE" ]; then
+    INSTALLED_NAME=`cat $INSTALL_FILE | grep $INSTALL_NAME`
+    [[ $INSTALLED_NAME == $INSTALL_NAME ]] && INSTALL_ALREADY=true
+fi
+
+if [ "$INSTALL_ALREADY" == true ]; then
+    echo "$INSTALL_NAME is already installed."
+else
+    echo "Installing $INSTALL_NAME."
+
+    curl https://baltocdn.com/helm/signing.asc | sudo apt-key add -
+    echo "deb https://baltocdn.com/helm/stable/debian/ all main" | sudo tee /etc/apt/sources.list.d/helm-stable-debian.list
+    sudo apt-get update
+    sudo apt-get install -y helm
+
+    [[ ! -d "$HOME/.config" ]] && mkdir $HOME/.config
+    echo $INSTALL_NAME >> $INSTALL_FILE
+fi
